requirements.txt
numpy
Flask
SQLAlchemy
pandas
matplotlib
seaborn
streamlit
pytesseract
Pillow
requests
ollama
python-dotenv